--- a/python/openvino_tokenizers/tokenizer_pipeline.py
+++ b/python/openvino_tokenizers/tokenizer_pipeline.py
@@ -348,6 +348,23 @@ class RegexSplitStep(PreTokenizatinStep):
                 f"got `{self.max_splits}`"
             )
 
+    def __add__(self, other: "RegexSplitStep") -> "RegexSplitStep":
+        if self.invert != other.invert:
+            raise ValueError("Cannot add two RegexSplitStep instances with different invert attributes")
+        if self.behaviour != other.behaviour:
+            raise ValueError("Cannot add two RegexSplitStep instances with different behaviour attributes")
+        if self.behaviour != "remove" or self.behaviour != "isolate":
+            raise ValueError('Only "remove" or "isolated" RegexSplit nodes can be merged')
+        if self.max_splits != other.max_splits:
+            raise ValueError("Cannot add two RegexSplitStep instances with different max_splits attributes")
+
+        return self.__class__(
+            split_pattern="|".join((self.split_pattern, other.split_pattern)),
+            invert=self.invert,
+            behaviour=self.behaviour,
+            max_splits=self.max_splits,
+        )
+
     @classmethod
     def split_by_chars(cls) -> "RegexSplitStep":
         return cls(split_pattern=".", invert=False, behaviour="isolate")
@@ -414,11 +431,11 @@ class RegexSplitStep(PreTokenizatinStep):
         )
 
     @classmethod
-    def digits_splitter(cls, behaviour="isolate") -> "RegexSplitStep":
+    def digits_splitter(cls, is_individual: bool = False) -> "RegexSplitStep":
         return cls(
-            r"\p{Nd}|\p{Nl}|\p{No}",
+            r"\p{Nd}|\p{Nl}|\p{No}" if is_individual else r"\p{Nd}+|\p{Nl}+|\p{No}+",
             invert=False,
-            behaviour=behaviour,
+            behaviour="isolate",
         )
 
     @classmethod
@@ -725,6 +742,7 @@ class BPETokenizationStep(TokenizationModelStep):
         else:
             special_tokens_outputs = []
 
+        # todo: check if this is still working after bytes-to-chars removal
         if special_tokens_outputs and pipeline.is_byte_level:
             special_tokens_outputs = pipeline.add_ragged_dimension(special_tokens_outputs)
             special_tokens_outputs = BytesToCharsStep().get_ov_subgraph(special_tokens_outputs)[-3:]
@@ -1457,6 +1475,37 @@ class TokenizerPipeline:
 
         self.steps = [step for step in self.steps if not isinstance(step, WhitespaceSplitStep)]
 
+    def merge_regex_split_steps(self) -> None:
+        if not any(isinstance(step, RegexSplitStep) for step in self.pre_tokenization_steps):
+            return
+
+        first_step_position = next(
+            idx for idx, step in enumerate(self.steps) if isinstance(step, RegexSplitStep)
+        )
+        steps_without_pre_tokenization = [step for step in self.steps if not isinstance(step, RegexSplitStep)]
+        old_regex_split_steps = [step for step in self.pre_tokenization_steps if isinstance(step, RegexSplitStep)]
+
+        new_regex_split_steps = []
+        while any(isinstance(step, RegexSplitStep) for step in old_regex_split_steps):
+            step_idx, current_step = next(
+                (idx, step) for idx, step in enumerate(old_regex_split_steps) if step is not None
+            )
+            old_regex_split_steps[step_idx] = None
+            new_regex_split_steps.append(current_step)
+
+            for idx, step in enumerate(old_regex_split_steps):
+                if step is None:
+                    continue
+
+                try:
+                    new_regex_split_steps[-1] = new_regex_split_steps[-1] + step
+                    old_regex_split_steps[idx] = None
+                except ValueError:
+                    pass
+
+        steps_without_pre_tokenization[first_step_position:first_step_position] = new_regex_split_steps
+        self.steps = steps_without_pre_tokenization
+
     def finalize(self) -> None:
         if self.finalized:
             return
@@ -1466,6 +1515,9 @@ class TokenizerPipeline:
 
         for step in copy(self.steps):
             step.finalize()
+
+        # merge after finalizing steps to make sure that BytesToCharsStep is removed
+        self.merge_regex_split_steps()
         self.finalized = True
 
     def get_tokenizer_ov_subgraph(self) -> Model: