--- a/tests/pass_rates.json
+++ b/tests/pass_rates.json
@@ -1,3 +1,3 @@
 {
-    "tests/tokenizers_test.py::test_": 0.95386559932589
+    "tests/tokenizers_test.py::test_": 0.9537835752857844
 }