Optimize tokenizers with multiple split steps by merging them into one step. Add DeepSeek-V3 as an example of such a tokenizer.

Benchmarks (i9-10980XE):
```bash
python benchmark.py <model_id> --dataset ShareGPT_V3_unfiltered_cleaned_split.json --seed 45 --tput
```

**deepseek-ai/DeepSeek-V3-0324**:
Before:
Sync:  OV: **1468.444 FPS**, HF: 1656.212 FPS, OV/HF: 0.8866282031381065
Async: OV: 7170.225 FPS, HF: 1656.212 FPS, OV/HF: 4.329292359446225
After:
Sync:  OV: **1618.003 FPS**, HF: 1676.049 FPS, OV/HF: 0.9653676397487092
Async: OV: 7178.811 FPS, HF: 1676.049 FPS, OV/HF: 4.28317497202293

**facebook/galactica-1.3b**:
Before:
Sync:  OV: **1366.586 FPS**, HF: 1325.043 FPS, OV/HF: 1.031351948327422
Async: OV: 7386.365 FPS, HF: 1325.043 FPS, OV/HF: 5.574432951671169
After:
Sync:  OV: **1613.592 FPS**, HF: 1351.466 FPS, OV/HF: 1.19395642793627
Async: OV: 7316.147 FPS, HF: 1351.466 FPS, OV/HF: 5.413488353161652